func for reading string, fixing
