--- alymova.ksenia/P4/create.cpp
@@ -1,10 +1,0 @@
-#include "dynamic_string.h"
-#include <cstdlib>
-#include <stdexcept>
-#include <cstddef>
-char* alymova::create(size_t size)
-{
-  char* str = reinterpret_cast< char* >(malloc((size + 1) * sizeof(char)));
-  str[size] = '\0';
-  return str;
-}
