+++ alymova.ksenia/P4/dynamic_string.h
@@ -4,7 +4,8 @@
 namespace alymova
 {
   char* create(size_t size);
-  char* increase_string(char* str, char* size_new);
+  char* copy_string(char* str, char* size_new);
   char* upper_string(char* str);
+  char* get_string(char* str, size_t& size, size_t& size_now, int ratio, char delim);
 }
 #endif
