+++ alymova.ksenia/P4/main.cpp
@@ -9,30 +9,17 @@ int main()
   char* str = alymova::create(size);
   if (str == nullptr)
   {
-    free(str);
     std::cerr << "Error: memory not allocate for string\n";
     return 1;
   }
-  char next = '\0';
   size_t size_now = 0;
+  char delim = '\n';
   std::cin >> std::noskipws;
-  while ((std::cin >> next) && (next != '\n'))
+  str = alymova::get_string(str, size, size_now, ratio, delim);
+  if (str == nullptr)
   {
-    str[size_now] = next;
-    size_now += 1;
-    if (size_now == size)
-    {
-      size *= ratio;
-      char* str_new = alymova::create(size);
-      if (str_new == nullptr)
-      {
-        free(str);
-        free(str_new);
-        std::cerr << "Error: memory not allocate for string\n";
-        return 1;
-      }
-      str = alymova::increase_string(str, str_new);
-    }
+    std::cerr << "Error: memory not allocate for string\n";
+    return 1;
   }
   if (size_now == 0)
   {
